refactor: split cpp and header files into separate folders
