+++ sim/include/index/componentType.h
