+++ sim/src/system/render.cpp
@@ -1,9 +1,9 @@
-#include "render.h"
+#include <system/render.h>
 #include <iostream>
 
-#include "../index/componentType.h"
-#include "../component/componentStoreExt.h"
-#include "../component/textureComponent.h"
+#include <index/componentType.h>
+#include <component/componentStoreExt.h>
+#include <component/textureComponent.h>
 #include <core/ics/componentStore.h>
 #include <core/ics/util/composable.h>
